NU-1858 fix properties in compare visibility
