+++ designer/client/src/components/modals/CompareVersionsDialog.tsx
@@ -22,6 +22,7 @@ import { Option, TypeSelect } from "../graph/node-modal/fragment-input-definitio
 import { WindowHeaderIconStyled } from "../graph/node-modal/nodeDetails/NodeDetailsStyled";
 import Icon from "../../assets/img/toolbarButtons/compare.svg";
 import i18next from "i18next";
+import { Properties } from "../properties/properties";
 
 const initState: State = {
     otherVersion: null,
@@ -181,7 +182,7 @@ const VersionsForm = ({ predefinedOtherVersion }: Props) => {
     };
 
     const printProperties = (property) => {
-        return property ? <NodeDetailsContent node={property} /> : <div className="notPresent">Properties not present</div>;
+        return property ? <Properties editedProperties={property} /> : <div className="notPresent">Properties not present</div>;
     };
 
     const versionOptions: Option[] = useMemo(() => {
